% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMACS FUNCTION.R
\name{dMACS}
\alias{dMACS}
\title{Computes dMACS}
\usage{
dMACS(fit.cfa)
}
\arguments{
\item{fit.cfa}{Lavaan output object with two groups and a single
factor.}
}
\value{
Returns dMACS for each item.
}
\description{
Computes dMACS
}
\examples{
dMACS
}
