% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inv_effects.R
\name{invariance_eff}
\alias{invariance_eff}
\title{Invariance Effect Sizes}
\usage{
invariance_eff(
  df,
  items,
  group,
  nodewidth = 0.01,
  intercept_fix = 1,
  lowerLV = -10,
  upperLV = 10,
  ...
)
}
\arguments{
\item{df}{Multi-group dataframe}

\item{items}{vector of items for the target construct}

\item{group}{string defining grouping variable}

\item{nodewidth}{Steps tested}

\item{intercept_fix}{Which item should have a fixed intercept defaults to the first item}

\item{lowerLV}{Lower range of latent variable tested}

\item{upperLV}{Upper range of latent variable tested}

\item{...}{Passes on to lavaan CFA functions}
}
\value{
Returns a dataframe with a row for each item and one column for each effect size
}
\description{
Invariance Effect Sizes
}
