% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavaan_helper.R
\name{lavTestScore.clean}
\alias{lavTestScore.clean}
\title{Get more comprehensible output from lavTestScore}
\usage{
lavTestScore.clean(lavaan.fit, ...)
}
\arguments{
\item{lavaan.fit}{Model fitted with lavaan}

\item{...}{Arguments passed to lavTestScore}
}
\description{
Get more comprehensible output from lavTestScore
}
\author{
Maksim Rudnev
}
