% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mg_rel_table.R
\name{mg_rel_table}
\alias{mg_rel_table}
\title{Multi-group reliability table}
\usage{
mg_rel_table(df_s, measure_list, group, digitn = 3, seed = 2711)
}
\arguments{
\item{df_s}{The full dataframe with all groups and items.}

\item{measure_list}{A named list of vectors containing the item names.
The format should be list(measure_name1 = c('Item1', 'Item2', 'Item3'), measure_name2 = c('Item1', 'Item2', 'Item3'))}

\item{group}{Grouping variable in the dataset as string for example "country"}

\item{digitn}{Controls the amount of digits shown in the output}

\item{seed}{Seed for the bootstrapped confidence intervals}
}
\value{
Returns a formatted dataframe with the reliability of all constructs by group
}
\description{
Multi-group reliability table
}
