% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pooled.R
\name{pancultural}
\alias{pancultural}
\title{Creating a Pan-Cultural Loading Matrix}
\usage{
pancultural(df, group, nfactors)
}
\arguments{
\item{df}{A data frame contains the variables for the exploratory factor
analysis and the grouping variable.}

\item{group}{The name of the column tht cointains the grouping supplied as a
string.}

\item{nfactors}{The number of factors expected.}
}
\value{
returns a Pan-Cultural loading matrix.
}
\description{
Creating a Pan-Cultural Loading Matrix
}
\examples{
pancultural(example, "country", 5)
}
