% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procrustes.R
\name{prost}
\alias{prost}
\title{Procrustes rotation function, returning Tucker's Phi}
\usage{
prost(loading, norm, rotated = FALSE, digits = 2)
}
\arguments{
\item{loading}{A correlation matrix to be rotated towards a target}

\item{norm}{A correlation matrix that is the goal of the rotation}

\item{rotated}{A TRUE/FALSE operator indicating if the rotated matrix should
be returned in addition to Tucker's Phi}

\item{digits}{The number of digits to be displayed in the output, defaults to 2}
}
\value{
Returns Tuckers Phi evaluating the congruence of the loading matrix
to the normative matrix
}
\description{
Procrustes rotation function, returning Tucker's Phi
}
