% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bonferroni_chi.R
\name{release_bonferroni}
\alias{release_bonferroni}
\title{Examening chisquare improvement if paths are unconstrained.
The function returns the paths to be unconstrained based on chisquare change.
Adjusted P-values are calculated based on iterative Bonferroni corrections.}
\usage{
release_bonferroni(lavaan.fit, exp_p = 0.05, ndigit = 3, ...)
}
\arguments{
\item{lavaan.fit}{Model fitted with lavaan}

\item{exp_p}{Expected p-value}

\item{ndigit}{Number of digits to round chi and p to}

\item{...}{Arguments passed to lavTestScore}
}
\description{
Examening chisquare improvement if paths are unconstrained.
The function returns the paths to be unconstrained based on chisquare change.
Adjusted P-values are calculated based on iterative Bonferroni corrections.
}
\author{
Maksim Rudnev
}
