% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_group.R
\name{splitgroup}
\alias{splitgroup}
\title{Split by groups}
\usage{
splitgroup(df, group, named = FALSE, name.list = NA)
}
\arguments{
\item{df}{Dataframe}

\item{group}{Variable from the dataset that defines the groups}

\item{named}{TRUE/FALSE argument wheter the resulting list should be named}

\item{name.list}{Supply a list of names same length as number of groups}
}
\value{
Returns a list of dataframes with only the selected groups
}
\description{
Split by groups
}
