% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_hat.R
\name{gamma_hat_scaled}
\alias{gamma_hat_scaled}
\title{Gamma Hat from MLM fitted lavaan object}
\usage{
gamma_hat_scaled(object)
}
\arguments{
\item{object}{A lavaan output object that was fitted with a MLM estimator}
}
\description{
Gamma Hat from MLM fitted lavaan object
}
