% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavaan_helper.R
\name{lavTestScore.clean}
\alias{lavTestScore.clean}
\title{Get more comprehensible output from lavTestScore}
\usage{
lavTestScore.clean(lavaan.fit, ndigit = 3, ...)
}
\arguments{
\item{lavaan.fit}{Model fitted with lavaan}

\item{ndigit}{Defines the rounding}

\item{...}{Arguments passed to lavTestScore}
}
\value{
Returns a dataframe which contains one row for each constrained parameter in the model together with a chi-square test indicating whether the parameter significantly differs between groups.
This is a cleaned version identical to \code{\link[lavaan:lavTestScore]{lavTestScore}}.
}
\description{
Get more comprehensible output from lavTestScore
}
\author{
Maksim Rudnev
}
